모니터링/requirements_pyqt6.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
decorator==5.1.1
Django==5.1
et-xmlfile==1.1.0
fonttools==4.53.0
frozenlist==1.4.1
idna==3.7
image==1.5.33
imutils==0.5.4
kiwisolver==1.4.5
matplotlib==3.9.0
multidict==6.1.0
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.5
openpyxl-image-loader==1.0.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
propcache==0.2.0
psutil==6.0.0
PyAudio==0.2.14
pycparser==2.22
pygame==2.5.2
pyparsing==3.1.2
pypng==0.20220715.0
pyqt-toast-notification==1.2.0
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.8.0
pyqtspinner==2.0.0
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyuac==0.0.3
pyzbar==0.1.9
qrcode==7.4.2
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
QtPy==2.4.1
rel==0.4.9.19
requests==2.32.1
shiboken6==6.7.2
six==1.16.0
sounddevice==0.4.7
SpeechRecognition==3.10.4
sqlparse==0.5.1
tee==0.0.3
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vulture==2.11
websocket-client==1.8.0
yarl==1.15.2
